fix circular deps, add todos
