+++ src/core/app_config/index.ts
@@ -1,41 +1,18 @@
-import { i18n } from '~core/localization';
-
-export interface AppConfig {
-  API_GATEWAY: string;
-  BOUNDARIES_API: string;
-  REPORTS_API: string;
-  BIVARIATE_TILES_RELATIVE_URL: string;
-  BIVARIATE_TILES_SERVER?: string;
-  BIVARIATE_TILES_INDICATORS_CLASS: string;
-  REFRESH_INTERVAL_SEC: number;
-  MAP_ACCESS_TOKEN: string;
-  MAP_BASE_STYLE: string;
-  LAYERS_BY_DEFAULT: string[];
-  KEYCLOAK_URL: string;
-  KEYCLOAK_REALM: string;
-  KEYCLOAK_CLIENT_ID: string;
-  YANDEX_METRICA_ID?: number[];
-  AUTOFOCUS_PADDINGS: [number, number, number, number];
-  AUTOFOCUS_ZOOM: number;
-  INTERCOM_DEFAULT_NAME?: string;
-  INTERCOM_APP_ID?: string;
-  INTERCOM_SELECTOR?: string;
-  FEATURES_BY_DEFAULT: string[];
-  DEFAULT_FEED: string;
-  OSM_EDITORS: OsmEditorConfig[];
-}
-
-export interface OsmEditorConfig {
-  id: string;
-  title: string;
-  url: string;
-}
+import type { AppConfig } from './types';
 
 declare global {
   interface Window {
     konturAppConfig: AppConfig;
   }
 }
+
+const DEFAULT_FEED_OBJECT = {
+  feed: 'kontur-public',
+  name: 'Kontur Public',
+  description:
+    'The feed contains real-time data about Cyclones, Droughts, Earthquakes, Floods, Volcanoes, Wildfires.',
+  default: true,
+};
 
 export default (() => {
   const konturAppConfig = globalThis.window.konturAppConfig ?? {};
@@ -52,7 +29,7 @@ export default (() => {
     layersByDefault: konturAppConfig.LAYERS_BY_DEFAULT,
     featuresByDefault: konturAppConfig.FEATURES_BY_DEFAULT,
     defaultFeed: konturAppConfig.DEFAULT_FEED,
-    defaultFeedObject: getDefaultFeedObject(konturAppConfig.DEFAULT_FEED),
+    defaultFeedObject: DEFAULT_FEED_OBJECT, // translation should occur later after i18n init, getDefaultFeedObject(konturAppConfig.DEFAULT_FEED),
     keycloakUrl: konturAppConfig.KEYCLOAK_URL,
     keycloakRealm: konturAppConfig.KEYCLOAK_REALM,
     keycloakClientId: konturAppConfig.KEYCLOAK_CLIENT_ID,
@@ -95,16 +72,4 @@ if (import.meta.env?.PROD) {
   `,
     'color: #bada55',
   );
-}
-
-function getDefaultFeedObject(feed?: string) {
-  // Change this solution when new default feed will be added
-  if (!feed || feed !== 'kontur-public')
-    console.warn('WARNING! Default feed provided via config is incorrect or absent');
-  return {
-    feed: 'kontur-public',
-    name: i18n.t('configs.Kontur_public_feed'),
-    description: i18n.t('configs.Kontur_public_feed_description'),
-    default: true,
-  };
 }
