+++ src/core/app_config/postinit.ts
@@ -1,0 +1,15 @@
+import { i18n } from '~core/localization';
+
+export function getDefaultFeedObject(feed?: string) {
+  // Change this solution when new default feed will be added
+  if (!feed || feed !== 'kontur-public')
+    console.warn('WARNING! Default feed provided via config is incorrect or absent');
+  return {
+    feed: 'kontur-public',
+    name: i18n.t('configs.Kontur_public_feed'),
+    description: i18n.t('configs.Kontur_public_feed_description'),
+    default: true,
+  };
+}
+
+// TODO: change DefaultFeedObject in appconfig after i18n initialization
