+++ src/core/localization/TranslationService.ts
@@ -1,13 +1,13 @@
 import i18n from 'i18next';
 import { initReactI18next } from 'react-i18next';
 import LanguageDetector from 'i18next-browser-languagedetector';
-import { currentUserAtom } from '~core/shared_state';
 import en_common from './translations/en/common.json';
 // As files are generated only before dev/build, so from start we don't have them
 // @ts-ignore
 import es_common from './translations/es/common.json';
 // @ts-ignore
 import ar_common from './translations/ar/common.json';
+
 
 i18n
   .use(LanguageDetector)
@@ -39,9 +39,3 @@ export const TranslationService = {
   },
   instance: i18n,
 };
-
-currentUserAtom.subscribe(({ language }) => {
-  i18n
-    .changeLanguage(language)
-    .catch((e) => console.warn(`Attempt to change language to ${language} failed`));
-});
